+++ Runtime/Scripts/Base/General.cs
@@ -2,54 +2,21 @@
  * File Name: General.cs
  * Description: Script that contains general utility functions
  * 
- * Authors: Will Lacey
+ * Author(s): Kokowolo, Will Lacey
  * Date Created: October 14, 2020
  * 
  * Additional Comments: 
  *      File Line Length: 120
- *
- *      This script relates to MathUtils.cs
  **/
 
+using System.Collections;
+using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.InputSystem;
-using UnityEngine.EventSystems;
-using System;
-using Diagnostics = System.Diagnostics;
 
 namespace Kokowolo.Utilities
 {
     public static class General
     {
-        #region Mouse Raytracing Functions
-
-        public static bool IsMouseOverUI()
-        {
-            return EventSystem.current.IsPointerOverGameObject();
-        }
-
-        [Obsolete("MouseScreenPointToRay is deprecated, please use MouseScreenPointToRaycastHit instead.")]
-        public static Ray MouseScreenPointToRay(Camera camera = null)
-        {
-            if (!camera) camera = Camera.main;
-            return camera.ScreenPointToRay(Mouse.current.position.ReadValue());
-        }
-
-        public static bool MouseScreenPointToRaycastHit(out RaycastHit hit, LayerMask layerMask, 
-            float maxDistance = Mathf.Infinity, bool debugDrawLine = false)
-        {
-            // Get MouseScreenPointToRay
-            Ray ray = Camera.main.ScreenPointToRay(Mouse.current.position.ReadValue());
-
-            bool valid = Physics.Raycast(ray, out hit, maxDistance, layerMask);
-            if (debugDrawLine) Debug.DrawLine(ray.origin, hit.point, Color.white, 1f);
-            return valid;
-        }
-
-        #endregion
-
-        #region Other Funtions
-
         public static T CacheGetComponent<T>(this MonoBehaviour monoBehaviour, ref T component) where T : Component
         {
             if (!component) component = monoBehaviour.GetComponent<T>();
@@ -88,7 +55,5 @@ namespace Kokowolo.Utilities
             
             return Application.persistentDataPath;
         }
-
-        #endregion
     }
 }